fix(route): swallowed routes due to nonunique map key
The path of routes was used as unique key in a map. However, the path
may be duplicated for multiple verbs/methods (GET/POST etc). As a result
a GET-route could be 'swallowed' by a POST-route.

This is fixed by combining the method and path to create a unique key.

Fixes: BartOtten/routex_example#1