feat(core): introduce Matchable module
Matchables are an essential part of Routex. They are used to match run time
routes with compile time routes.

This module provides functions to create Matchables, convert them to match
pattern AST as well as function heads AST and to check if the routing values
of two Matchable records match.